requirements.txt
fastapi
folium
pycountry
python-dotenv
requests
uvicorn
